Enlargen page area even more
The previous setting DIV=11 leads to already very long lines
(~83 characters/line). However, the text area of the pages is
somewhat too small for this value, as I easily filled 5 pages
with the notational introduction.

There are various options to increase the amount of information
per page:
1. Decrease font size
2. Decrease line spacing
3. Decrease page margins

I checked with a ruler in GIMP and 1. and 2. are not viable
approaches, as the font size is exactly 12pt (measured) and the
line spacing is exact 1.5, both of which are required by the PO
(PhD regulations of the University of Stuttgart). The only option
that remains is number 3, and this is achieved by incrementing
the DIV value by 1 (DIV=12). The lines are even longer now
(~90 characters/line) and one additional line fits on the page
(33 lines/page).

I know 90 chars/line would be too high for a novel for example,
but it's the only way. The PO doesn't make any requirements here,
only that the page margins have to be at least 2cm on either side.