Swap flip books for correct order
On odd pages, the flip book has to start from the back, while on
even pages, it has to start from the front. The SG animation
can be reversed, so we put that on the odd pages. Otherwise, the
B-splines would have to start with degree 3 and then go down
to 1... Plus, we don't have to rename files, as the SG animation
can be played in reverse as well.