Add Jamil13Literature to old/
Sorry, I refuse to cite Jamil13Literature directly. It's a massive
compilation of test problems and all, but it just contains so many
errors. It's not just a few---there are errors literally
everywhere. The domain might be wrong, the sign of the objective
function might be wrong, a "+" might be missing, and so on...
I've lost confidence in that source.

I find it more appropriate to invest the time to cite the original
literature instead.