Fix very weird layout bug for draft mode = false
As far as I could check, it's as follows: If no page with a page
number on it (i.e., `empty` page style) appears before the table
of contents, then the layout for rest of the thesis is completely
different (at least for the introduction), meaning that words
wrap differently, despite exactly the same page area settings.

Consequently, when the draft mode was disabled for the first time
since quite a while, this bug happened, because the progress
table (which is hidden if the draft mode is disabled) had the
only two pages with numbers on it before the table of contents.

I have absolutely no idea why this happens, and at this stage,
I really can't be bothered to investigate further. So I'm fixing
this by just showing page numbers on the dedication.